New Additions
Added sqlite3 database, created LIBRARY table. Need to pass JSON data
back to the server.js so that it can process the JSON to build an AJAX
response. The AJAX responses are to build the html pages "library",
"home stations", etc.

Run a count() query on the database to change the number of songs in
library.

Add a column to LIBRARY table (file_location).