Rooms
This branch should follow the following notes:

USER's ROOM
1) First, implement functionality so that EVERYONE using the local
server (we call this the USER's ROOM) can see who is listening to what
and what current time they are at. So, setup the USER's ROOM
functionality. When people join a USER's room, they can see what
everyone is listening to, its all being updated.

when a new user connects: a new user box pops up on my main screen
when a user starts listening to a song: the album artwork displays on
that box and the track progress is displayed.
when a user pauses a song: it should be displayed as well
when a user disconnects: the box should dissappear

2) Second, implement the functionality to JOIN listening with a user.
- This will require the use of Socket.IO rooms.
- Perhaps, everytime the user starts listening to music, a
listening-session socket is piggybacked alongside it. So that when
someone else wants to join, they can just start receiving the track
position data from that listening-session.
- This will also require knowing how to start a Web API audio track from
a certain position ( noteOn() ) ?