Additional Client Socket Logic
-Added more client side socket logic
-Currently, every socket in the default room receives updates about the
current track position, track name, album, artist and key (to be used to
retrieve the album-artwork) from the file location in the database. This
is still yet to be implemented but we should think about doing this in a
more efficient way.

-Changes to the layout have been made on the Index.html page. We want to
follow with this structure. Media Queries will need to be done to handle
displaying it on smaller devices.