New Views
Been working on new views, a lot of refractoring is going to have to be
done on the project. Currently researching the best ways to go about a
few things.

Each view will need to be refractored to .ejs templates

For the Library.html view, some of the sidebar animation is working.